+++ R/get_satellite_imagery.R
@@ -1,5 +1,5 @@
 
-# get_available_imagery() -------------------------------------------------------
+# get_available_imagery() ------------------------------------------------------
 
 #' Get a Listing of Available BoM Satellite GeoTIFF Imagery
 #'
@@ -87,7 +87,7 @@ get_available_imagery <- function(product_id = "all") {
 #' @param cache Logical.  Store image files locally for later use?  If
 #' \code{FALSE}, the downloaded files are removed when R session is closed. To
 #' take advantage of cached files in future sessions, use \code{cache = TRUE}.
-#' Defaults to \code{FALSE}. Value is optional.
+#' Defaults to \code{FALSE}.  Value is optional.
 #'
 #' @details Valid BoM satellite Product IDs include:
 #'\describe{